requirements.txt
streamlit
yt-dlp
git+https://github.com/openai/whisper.git
google-generativeai
python-dotenv
